Upgrade ensure_cube_aggregatable_across_cases to public operator
